jingtai

+++ interfaces/kits/native/want/src/want.cpp
@@ -2251,7 +2251,7 @@ void Want::SetEntities(const std::vector<std::string> &entities)
 
 int32_t Want::Flags_ConvertEts2Native(const int32_t index)
 {
-    if (index < 0 || index >= FlagsArray_.size()) {
+    if (index < 0 || index >= (int32_t)FlagsArray_.size()) {
         ABILITYBASE_LOGE("Flags_ConvertEts2Native failed index:%{public}d", index);
         return 0;
     }
@@ -2260,7 +2260,7 @@ int32_t Want::Flags_ConvertEts2Native(const int32_t index)
 
 int32_t Want::Flags_ConvertNative2Ets(const int32_t nativeValue)
 {
-    for (int index = 0; index < FlagsArray_.size(); index++) {
+    for (int index = 0; index < (int)FlagsArray_.size(); index++) {
         if (nativeValue == FlagsArray_[index]) {
             return index;
         }
@@ -2271,7 +2271,7 @@ int32_t Want::Flags_ConvertNative2Ets(const int32_t nativeValue)
 
 std::string Want::Action_ConvertEts2Native(const int32_t index)
 {
-    if (index < 0 || index >= ActionArray_.size()) {
+    if (index < 0 || index >= (int32_t)ActionArray_.size()) {
         ABILITYBASE_LOGE("Action_ConvertEts2Native failed index:%{public}d", index);
         return "";
     }
@@ -2280,7 +2280,7 @@ std::string Want::Action_ConvertEts2Native(const int32_t index)
 
 int32_t Want::Action_ConvertNative2Ets(const std::string nativeValue)
 {
-    for (int32_t index = 0; index < ActionArray_.size(); index++) {
+    for (int32_t index = 0; index < (int32_t)ActionArray_.size(); index++) {
         if (nativeValue == ActionArray_[index]) {
             return index;
         }
